SIL: Move `SerializedSILLoader.h` to the `include/swift/SIL` directory.
The `SerializedSILLoader` class is implemented in SIL, not Serialization, so the previous location of the header was misleading.
